refactor vault usage
