+++ packages/mina-wallet/src/types.ts
@@ -1,4 +1,3 @@
-import { Network } from '@palladxyz/key-generator'
 import {
   AccountInfo,
   Mina,
@@ -14,20 +13,18 @@ import { PublicCredential } from '@palladxyz/vault'
 export interface MinaWallet {
   createWallet(
     walletName: string,
-    network: Network,
     accountNumber: number
   ): Promise<{ publicKey: string; mnemonic: string } | null>
 
   restoreWallet(
     walletName: string,
-    network: Network,
     mnemonic: string,
     accountNumber: number
   ): Promise<{ publicKey: string } | null>
 
   getCurrentWallet(): PublicCredential | null
 
-  getAccounts(): string[]
+  //getAccounts(): string[]
 
   getAccountInfo(publicKey: Mina.PublicKey): Promise<AccountInfo>
